+++ es-design-system/components/ds-molecules-list.vue
@@ -31,6 +31,11 @@
             </b-link>
         </li>
         <li>
+            <b-link to="/molecules/es-file-input">
+                EsFileInput
+            </b-link>
+        </li>
+        <li>
             <b-link to="/molecules/es-footer">
                 EsFooter
             </b-link>
