+++ es-design-system/components/ds-organisms-list.vue
@@ -1,11 +1,6 @@
 <template>
     <ul>
         <li>
-            <b-link to="/organisms/es-file-upload">
-                EsFileUpload
-            </b-link>
-        </li>
-        <li>
             <b-link to="/organisms/es-form">
                 EsForm
             </b-link>
